test(cli): extract test suite for checking loopback project directory
